cmake versiones

+++ Rodrigo/imgtxtenh/src/imgio_magick.c
@@ -151,9 +151,9 @@ int getalpha_magick_graym( Img* img, gray* gimg ) {
   int n;
   for( n=img->width*img->height-1; n>=0; n-- )
 #if MAGICKCORE_QUANTUM_DEPTH == 16
-    gimg[n] = GetPixelAlpha(pixs+n) >> 8;
+    gimg[n] = (int)GetPixelAlpha(pixs+n) >> 8;
 #elif MAGICKCORE_QUANTUM_DEPTH == 8
-    gimg[n] = GetPixelAlpha(pixs+n);
+    gimg[n] = (int)GetPixelAlpha(pixs+n);
 #endif
 
   return SUCCESS;
@@ -171,9 +171,9 @@ int getpixels_magick_graym( Img* img, gray* gimg ) {
   int n;
   for( n=img->width*img->height-1; n>=0; n-- )
 #if MAGICKCORE_QUANTUM_DEPTH == 16
-    gimg[n] = GetPixelGray(pixs+n) >> 8;
+    gimg[n] = (int)GetPixelGray(pixs+n) >> 8;
 #elif MAGICKCORE_QUANTUM_DEPTH == 8
-    gimg[n] = GetPixelGray(pixs+n);
+    gimg[n] = (int)GetPixelGray(pixs+n);
 #endif
 
   return SUCCESS;
@@ -191,13 +191,13 @@ int getpixels_magick_pixelm( Img* img, pixel* cimg ) {
   int n;
   for( n=img->width*img->height-1; n>=0; n-- ) {
 #if MAGICKCORE_QUANTUM_DEPTH == 16
-    cimg[n].r = GetPixelRed(pixs+n) >> 8;
-    cimg[n].g = GetPixelGreen(pixs+n) >> 8;
-    cimg[n].b = GetPixelBlue(pixs+n) >> 8;
+    cimg[n].r = (int)GetPixelRed(pixs+n) >> 8;
+    cimg[n].g = (int)GetPixelGreen(pixs+n) >> 8;
+    cimg[n].b = (int)GetPixelBlue(pixs+n) >> 8;
 #elif MAGICKCORE_QUANTUM_DEPTH == 8
-    cimg[n].r = GetPixelRed(pixs+n);
-    cimg[n].g = GetPixelGreen(pixs+n);
-    cimg[n].b = GetPixelBlue(pixs+n);
+    cimg[n].r = (int)GetPixelRed(pixs+n);
+    cimg[n].g = (int)GetPixelGreen(pixs+n);
+    cimg[n].b = (int)GetPixelBlue(pixs+n);
 #endif
   }
 
@@ -293,14 +293,14 @@ int togray_magick( Img* img ) {
   for( n=img->width*img->height-1; n>=0; n-- )
 #if MAGICKCORE_QUANTUM_DEPTH == 16
     gimg[n] = (gray)( 0.5 +
-              0.299*(GetPixelRed(pixs+n)>>8) +
-              0.587*(GetPixelGreen(pixs+n)>>8) +
-              0.114*(GetPixelBlue(pixs+n)>>8) );
+              0.299*((int)GetPixelRed(pixs+n)>>8) +
+              0.587*((int)GetPixelGreen(pixs+n)>>8) +
+              0.114*((int)GetPixelBlue(pixs+n)>>8) );
 #elif MAGICKCORE_QUANTUM_DEPTH == 8
     gimg[n] = (gray)( 0.5 +
-              0.299*(GetPixelRed(pixs+n)) +
-              0.587*(GetPixelGreen(pixs+n)) +
-              0.114*(GetPixelBlue(pixs+n)) );
+              0.299*((int)GetPixelRed(pixs+n)) +
+              0.587*((int)GetPixelGreen(pixs+n)) +
+              0.114*((int)GetPixelBlue(pixs+n)) );
 #endif
 
   int err = setpixels_magick_graym( img, gimg );
